+++ test/unit/fuzz-tools.test.js
@@ -281,6 +281,7 @@ import * as drSeverityV301 from '../../lib/dr-severity-v301.js';
 import * as isisV302 from '../../lib/isis-v302.js';
 import * as anaphylaxisV303 from '../../lib/anaphylaxis-v303.js';
 import * as dstV304 from '../../lib/dst-v304.js';
+import * as crsV305 from '../../lib/crs-v305.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -507,6 +508,7 @@ const MODULES = {
   'isis-v302.js': isisV302,
   'anaphylaxis-v303.js': anaphylaxisV303,
   'dst-v304.js': dstV304,
+  'crs-v305.js': crsV305,
   'tb-testing.js': tbTesting,
 };